requirements.txt
ffmpeg
gym==0.21.0
gym-retro
gymnasium[accept-rom-license]
moviepy
optuna
stable-baselines3[extra] 
opencv-python
optuna
torch==1.10.1+cu113 
torchvision==0.11.2+cu113 
torchaudio===0.10.1+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html